[ose][downstream][media][ls2] Match URIs array with URITypes
:Release Notes:
Match URIs array with URITypes

:Detailed Notes:
luna_service_uris array is indexed based on URIType, but there was
a mismatch between the sizes of each.
As some services are not available in OSE, use empty strings so that
there's no buffer overrun when accessing it based on any URIType.

:Testing Performed:
[QO-1043] Basic testing of Blink BAT for Open on the RPI target

:QA Notes:
Test-cases: [QO-1043]

:Issues Addressed:
[PLAT-80054] Port "[ose][downstream][media][ls2] Match URIs array with
URITypes"
[PLAT-72650] [OSE] [chromium68] Static analysis issues

Change-Id: I937c49f3b970849e517999825b5118691e234807
Originally-Reviewed-on: http://gpro.lge.com/#/c/228141/
Reviewed-on: http://gpro.lge.com/244753