[open][neva_upstreamable][app_shell] Do not enable LimitedMediaActivationCountPolicy by default
:Release Notes:
Do not enable LimitedMediaActivationCountPolicy by default

:Detailed Notes:
It is decided to activate LimitedMediaActivationCountPolicy only on
some devices. On chromium68, we were activating it by bb files by
adding a chromium runtime flag '--max-activated-media-players=1'.
But on chromium72, it was added on run_app_shell script, which
activates LimitedMediaActivationCountPolicy on all environments.
This patch removes the flag from run_app_shell
script. '--max-activated-media-players' will be added by sed commend
on bb files as it was on chromium68.

:Testing Performed:
Tested QO-2324 on raspberrypi3 board, with build.

:QA Notes:
Test-cases: QO-2324

:Issues Addressed:
[PLAT-83375] [WebOSE]-video is not playing in facebook.

Change-Id: Ib7004bc4fbd13bd979c7b590399f6589a757e347
Originally-Reviewed-on: http://gpro.lge.com/250926
Reviewed-on: http://gpro.lge.com/250960
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>