[open][neva_upstreamable][ozone][webos] Unset mouse move handler on webOSMouse leave
:Release Notes:
Do not miss spotlight focusing after mouse cursor is hidden.

:Detailed Notes:
There were some special cases where the webOSMouse enter event was not
happening after webOSMouse leave event (that happens when mouse cursor
is dismissed). Because of that, if an application would relay on that
instead of cursor visibility events for spotlight focus, then it would
stop working easily after first focus hide.

To avoid that, on webOS we explicitely unset the focus when the cursor
is hidden, as in webOS when focus is back it will be reset to the center
of the screen.

:Testing Performed:
1. Launch player app.
2. Mouse hover and select several options in left panel.
3. Let mouse cursor go away.
4. Wave mouse and see if spotlight works.

:QA Notes:
Test-cases: QS-8538, QS-6687, QS-6690

:Issues Addressed:
[PLAT-85210] When hover the USB mouse cursor, unable to focus the settings UI.
[PLAT-85369] Spotlight not working on mouse hover after mouse cursor dismissed

Change-Id: I667ac4422c51c618f6555f86767c46d75a4e5e89
Originally-Reviewed-on: http://gpro.lge.com/255039
Reviewed-on: http://gpro.lge.com/255051
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>