[open][neva_upstreamable][a11y] Load webOS accessibility l10n resources
:Release Notes:
Load webOS accessibility strings translations

:Detailed Notes:
webOS acccessibility support was not loading the translation strings
required for accessibility, nor changing them on changing language.

First problem was that the resource bundles were not including the
translations in components/accessibility. This change adds them
to the installed and loaded resource bundles.

Second problem was that we were not updating the resource bundles
with locales in browser process (we did in renderer process though).
So added in WebOSPlatform hook for system locale change code to update
the bundle when locale is changed.

:Testing Performed:
1. Go to settings
2. Go to Ipv6 enabled checkbox and click on it. Check the "checked"
and "unchecked" states are sent to TTS (with the currently selected
menu language).
3. Change menus language to a different language.
4. Try again in Ipv6 enabled checkbox if the "checked" and "unchecked"
strings are translated.

:QA Notes:
Test-cases: GT-27897

:Issues Addressed:
[PLAT-86030] [Accessibility] Not changed Locale

Change-Id: Iee8fabc548a9623a8841fe20e081f7bd7810e14d
Originally-Reviewed-on: http://gpro.lge.com/256815