fixup! [open][downstream][disk_cache][emmc_opt] Optimize HTTP cache for eMMC
:Release Notes:
Set enable_emmc_optimizations to false as default

:Detailed Notes:
enable_emmc_optimizations is to reduce writes to eMMC for http cache.
In details, please refer to the related commit http://gpro.lge.com/245102.
As the commit speficies the NOTE, it does break HTML5 application cache
feature. The problem can be observed in PLAT-84800.
So this patch sets enable_emmc_optimizations to false as default.

:Testing Performed:
1. Loading phraffle.com
2. Loading daum.net
3. Disconnect LAN cable
4. Loading phraffle.com and check whether it is loaded successfully.

:QA Notes:
Test-cases: QWA-127

:Issues Addressed:
[PLAT-84800] Black screen is displaying
[PLAT-86669] [webOS OSE 2.0]: Black screen appears for http://phraffle.com

Change-Id: I5c9733abcee4278e0b51c1fad0169dcd395237f4
Reviewed-on: http://gpro.lge.com/258361
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>