fixup! [ose][ozone][wayland] Implement multiple touch event handling
:Release Notes:
Implement multiple touch event handling

:Detailed Notes:
According to LSM changes:
http://gpro.lge.com/#/c/260746/

1. We should separate by device id for
grabbing window handling by touch press.

2. Correct file names for previously added are changed.

:Testing Performed:

We are observing an issue with actuation when touching elements.
This causes a problem when we are attempting to scroll on both screens at the same time.
Interaction with one element prevents scroll with the other element,
and if you keep your finger in place you can utilize scroll on another screen (B) from screen (A)

When a user control scrolling on the screen A and screen B,
the scroll should be controlled on the screen A and screen B respectively.

The issue should be fixed after LSM issue - [PLAT-88675] will fixed
1. Load a page with scroll on display#0, #1
2. Start scroll on display#0 and do scrolling continuously.
3. Start scroll on display#1 -> Scroll works on display#0, display#1 independently
4. Stop scroll on display#1
Scroll still works on display#0, Scroll is stopped on display#1

:QA Notes:
Test-cases: QWA-1561
Depends-on: http://gpro.lge.com/#/c/260746/

:Issues Addressed:
[PLAT-87480] Migrate multiple VKB patch to webOS OSE

Change-Id: I379294b53c450678f8c82a4d7a9c4da5a721d835
Originally-Reviewed-on: http://gpro.lge.com/260210
Reviewed-on: http://gpro.lge.com/c/webosose/chromium72/+/263703
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>