[ose][neva_upstreamable][ozone][wayland] Set the maximum input length to VKB
:Release Notes:
Implemented a way for setting the maximum input length to VKB

:Detailed Notes:
For VKB UI feature:
The Enter button will be disabled if the input field
entered the max allowed characters.
The app will set this max length of the input field and
the imemanger expected to receive from
inputMethodHost()->maxTextLength().
The Enter button will be enabled if the character length is
less than Max input length defined by the apps.

For the feature, in the patch, we get the maximum length
from HTML element and set it to VKB
via text model Wayland interface.

:Testing Performed:
RSPi4 webosose build.
Check Wayland log files and check that
max input length value is set to VKB via text model Wayland interface.

:QA Notes:
Test-cases: QWA-1847

:Issues Addressed:
[PLAT-97821] [OSE][Auto] Migrate "Set the maximum input length to VKB"
patch
[PLAT-89649] Disable the Enter button after entering max characters

Change-Id: I34565452181d290107e22d840949c90426141625
Originally-Reviewed-on:
http://gpro.lge.com/c/webosose/chromium72/+/263731
Reviewed-on: http://gpro.lge.com/c/webosose/chromium72/+/265933
Reviewed-by: OSE Helper Bot <ose@lge.com>
Reviewed-by: Alexander Trofimov <alexander.trofimov@lge.com>
Reviewed-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>
Tested-by: Igor Makarov <igor.makarov@lge.com>
Tested-by: Lokesh Kumar Goel <lokeshkumar.goel@lge.com>